+++ packages/client/src/pages/game-menu/index.tsx
@@ -2,7 +2,7 @@ import { useState } from 'react'
 import { Countdown } from '../../components/count-down'
 import { GameEnd } from '../../components/game-end/game-end'
 import { Container, Button } from '@mui/material'
-import { Game } from '../_game'
+import { Game } from '../game'
 import { useNavigate } from 'react-router-dom'
 import { RouterName } from '../../router/types'
 import { useDispatch } from 'react-redux'
