+++ packages/client/src/pages/game/index.tsx
