+++ packages/client/src/pages/index.ts
@@ -5,6 +5,6 @@ export * from './sing-up'
 export * from './user-profile'
 export * from './settings'
 export * from './page-404'
-export * from './_game'
+export * from './game'
 export * from './page-500'
 export * from './game-menu'
