+++ packages/client/src/pages/sing-up/sing-up.tsx
@@ -5,7 +5,7 @@ import { Link, useNavigate } from 'react-router-dom'
 import { singUpInitialValues } from '../../constants/initialValues'
 import { singUpValidationSchema } from '../../constants/validationSchema'
 import { SignUpProps } from '../../services'
-import styles from './SingUp.module.css'
+import styles from './sing-up.module.css'
 import { RouterName } from '../../router/types'
 import { useState } from 'react'
 import { useSignUpMutation } from '../../store/api'
